+++ src/Twilio/Rest/Preview/TrustedComms.php
@@ -13,6 +13,7 @@ use Twilio\Domain;
 use Twilio\Exceptions\TwilioException;
 use Twilio\InstanceContext;
 use Twilio\Rest\Preview\TrustedComms\BrandedCallList;
+use Twilio\Rest\Preview\TrustedComms\BrandedChannelList;
 use Twilio\Rest\Preview\TrustedComms\BrandsInformationList;
 use Twilio\Rest\Preview\TrustedComms\BusinessList;
 use Twilio\Rest\Preview\TrustedComms\CpsList;
@@ -22,15 +23,18 @@ use Twilio\Version;
 
 /**
  * @property BrandedCallList $brandedCalls
+ * @property BrandedChannelList $brandedChannels
  * @property BrandsInformationList $brandsInformation
  * @property BusinessList $businesses
  * @property CpsList $cps
  * @property CurrentCallList $currentCalls
  * @property PhoneCallList $phoneCalls
+ * @method \Twilio\Rest\Preview\TrustedComms\BrandedChannelContext brandedChannels(string $sid)
  * @method \Twilio\Rest\Preview\TrustedComms\BusinessContext businesses(string $sid)
  */
 class TrustedComms extends Version {
     protected $_brandedCalls;
+    protected $_brandedChannels;
     protected $_brandsInformation;
     protected $_businesses;
     protected $_cps;
@@ -52,6 +56,13 @@ class TrustedComms extends Version {
             $this->_brandedCalls = new BrandedCallList($this);
         }
         return $this->_brandedCalls;
+    }
+
+    protected function getBrandedChannels(): BrandedChannelList {
+        if (!$this->_brandedChannels) {
+            $this->_brandedChannels = new BrandedChannelList($this);
+        }
+        return $this->_brandedChannels;
     }
 
     protected function getBrandsInformation(): BrandsInformationList {