+++ src/Twilio/Rest/Verify/V2/ServiceList.php
@@ -53,7 +53,9 @@ class ServiceList extends ListResource {
             'Psd2Enabled' => Serialize::booleanToString($options['psd2Enabled']),
             'DoNotShareWarningEnabled' => Serialize::booleanToString($options['doNotShareWarningEnabled']),
             'CustomCodeEnabled' => Serialize::booleanToString($options['customCodeEnabled']),
-            'Push' => Serialize::jsonObject($options['push']),
+            'Push.IncludeDate' => Serialize::booleanToString($options['pushIncludeDate']),
+            'Push.ApnCredentialSid' => $options['pushApnCredentialSid'],
+            'Push.FcmCredentialSid' => $options['pushFcmCredentialSid'],
         ]);
 
         $payload = $this->version->create('POST', $this->uri, [], $data);